Add Device.QuietOn.
This is useful since the LIFX LAN API requires a device to be powered on
for transitions to take effect, but automation often doesn't know (and
doesn't want to know) the current state in order to get to a
well-defined new state smoothly. QuietOn ensures the device will be in a
state for methods like SetColor and SetWaveform to work as expected.